rename old player controller, create new WIP player controller
